+++ fixed-vector/Data/Vector/Fixed/Strict.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE MagicHash            #-}
+{-# LANGUAGE UnboxedTuples        #-}
 {-# LANGUAGE UndecidableInstances #-}
 -- |
 -- Strict boxed vector which could hold any value. For lazy variant see
@@ -11,9 +12,10 @@ import Data.Primitive.SmallArray
 import Data.Monoid          (Monoid(..))
 import Data.Semigroup       (Semigroup(..))
 import Data.Data
+import Data.Primitive.Types (Prim)
 import qualified Data.Foldable    as F
 import qualified Data.Traversable as T
-import Foreign.Storable (Storable(..))
+import Foreign.Storable (Storable)
 import GHC.TypeLits
 import GHC.Exts (proxy#)
 import Prelude ( Show(..),Eq(..),Ord(..),Functor(..),Monad(..)
@@ -72,6 +74,8 @@ deriving via ViaFixed (Vec n) a instance (Arity n, NFData    a) => NFData    (Ve
 deriving via ViaFixed (Vec n) a instance (Arity n, Semigroup a) => Semigroup (Vec n a)
 deriving via ViaFixed (Vec n) a instance (Arity n, Monoid    a) => Monoid    (Vec n a)
 deriving via ViaFixed (Vec n) a instance (Arity n, Storable  a) => Storable  (Vec n a)
+-- | @since 2.0.1.0
+deriving via ViaFixed (Vec n) a instance (Arity n, Prim      a) => Prim      (Vec n a)
 
 instance (Arity n) => MVector (MVec n) a where
   basicNew =